src/modules/72-templates-library/components/TemplateDetails/__tests__/TemplateDetails.test.tsx
/*
 * Copyright 2021 Harness Inc. All rights reserved.
 * Use of this source code is governed by the PolyForm Shield 1.0.0 license
 * that can be found in the licenses directory at the root of this repository, also available at
 * https://polyformproject.org/wp-content/uploads/2020/06/PolyForm-Shield-1.0.0.txt.
 */

import React from 'react'
import { act, fireEvent, render } from '@testing-library/react'
import { useLocation } from 'react-router-dom'
import { defaultTo, unset } from 'lodash-es'
import produce from 'immer'
import { TestWrapper } from '@common/utils/testUtils'
import { mockTemplates, mockTemplatesSuccessResponse } from '@templates-library/TemplatesTestHelper'
import templateFactory from '@templates-library/components/Templates/TemplatesFactory'
import { mockTemplatesInputYaml } from '@pipeline/components/PipelineStudio/PipelineStudioTestHelper'
import { StepTemplate } from '@templates-library/components/Templates/StepTemplate/StepTemplate'
import * as commonHooks from '@common/hooks'
import { templatePathProps } from '@common/utils/routeUtils'
import routes from '@common/RouteDefinitions'
import { mockBranches } from '@gitsync/components/GitSyncForm/__tests__/mockdata'
import { TemplateDetails, TemplateDetailsProps } from '../TemplateDetails'

const gitAppStoreValues = {
  isGitSyncEnabled: false,
  isGitSimplificationEnabled: true,
  supportingGitSimplification: true,
  gitSyncEnabledOnlyForFF: false,
  supportingTemplatesGitx: true
}

const TEST_PATH = routes.toTemplateStudio(templatePathProps)
const TEST_PATH_PARAMS = {
  templateIdentifier: '-1',
  accountId: 'accountId',
  orgIdentifier: 'default',
  projectIdentifier: 'projectId',
  module: 'cd',
  templateType: 'Step'
}

const useGetTemplateMock = jest.fn()

jest.mock('@templates-library/components/TemplateInputs/TemplateInputs', () => ({
  ...jest.requireActual('@templates-library/components/TemplateInputs/TemplateInputs'),
  TemplateInputs: () => {
    return <div className="template-inputs-mock"></div>
  }
}))

jest.mock('services/template-ng', () => ({
  ...jest.requireActual('services/template-ng'),
  useGetTemplate: jest.fn().mockImplementation((...args) => {
    useGetTemplateMock(...args)
    return {}
  }),
  useGetTemplateInputSetYaml: jest
    .fn()
    .mockImplementation(() => ({ data: mockTemplatesInputYaml, refetch: jest.fn(), error: null, loading: false })),
  useListTemplateUsage: () => ({
    loading: false,
    data: {},
    refetch: jest.fn()
  })
}))

const fetchBranches = jest.fn(() => Promise.resolve(mockBranches))
jest.mock('services/cd-ng', () => ({
  useGetListOfBranchesByRefConnectorV2: jest.fn().mockImplementation(() => {
    return { data: mockBranches, refetch: fetchBranches }
  }),
  useListAllEntityUsageByFqn: () => ({
    loading: false,
    data: {},
    refetch: jest.fn()
  })
}))

function ComponentWrapper(props: TemplateDetailsProps): React.ReactElement {
  const location = useLocation()
  return (
    <React.Fragment>
      <TemplateDetails {...props} />
      <div data-testid="location">{`${location.pathname}${location.search}`}</div>
    </React.Fragment>
  )
}

describe('<TemplateDetails /> git experience', () => {
  afterEach(() => {
    useGetTemplateMock.mockReset()
  })
  beforeEach(() => {
    jest
      .spyOn(commonHooks, 'useMutateAsGet')
      .mockImplementation(jest.fn().mockReturnValue(mockTemplatesSuccessResponse))
  })

  test('Template GET API sends parent entity context in query params only when default behaviour is there', () => {
    const baseProps: TemplateDetailsProps = {
      template: defaultTo(mockTemplates?.data?.content?.[0], {}),
      storeMetadata: {
        connectorRef: 'connectorRefTest',
        storeType: 'REMOTE',
        branch: 'branchTest',
        repoName: 'manju-test-template-qq-12344'
      }
    }

    render(
      <TestWrapper path={TEST_PATH} pathParams={TEST_PATH_PARAMS} defaultAppStoreValues={gitAppStoreValues}>
        <ComponentWrapper {...baseProps} />
      </TestWrapper>
    )

    expect(useGetTemplateMock).toHaveBeenCalledWith({
      lazy: true,
      queryParams: {
        accountIdentifier: 'kmpySmUISimoRrJL6NL73w',
        branch: 'branchTest',
        getDefaultFromOtherRepo: true,
        orgIdentifier: 'default',
        parentEntityAccountIdentifier: 'accountId',
        parentEntityConnectorRef: 'connectorRefTest',
        parentEntityOrgIdentifier: 'default',
        parentEntityProjectIdentifier: 'projectId',
        parentEntityRepoName: 'manju-test-template-qq-12344',
        projectIdentifier: 'Templateproject',
        versionLabel: 'v4',
        repoIdentifier: undefined
      },
      requestOptions: {
        headers: { 'Load-From-Cache': 'true' }
      },
      templateIdentifier: 'manjutesttemplate'
    })
  })

  test('Template GET API doesnt send parent entity context in query params for inline templates', () => {
    const baseProps: TemplateDetailsProps = {
      template: defaultTo(mockTemplates?.data?.content?.[0], {}),
      storeMetadata: undefined
    }

    render(
      <TestWrapper path={TEST_PATH} pathParams={TEST_PATH_PARAMS} defaultAppStoreValues={gitAppStoreValues}>
        <ComponentWrapper {...baseProps} />
      </TestWrapper>
    )

    expect(useGetTemplateMock).toHaveBeenCalledWith({
      lazy: true,
      queryParams: {
        accountIdentifier: 'kmpySmUISimoRrJL6NL73w',
        getDefaultFromOtherRepo: true,
        orgIdentifier: 'default',
        projectIdentifier: 'Templateproject',
        versionLabel: 'v4'
      },
      requestOptions: {
        headers: { 'Load-From-Cache': 'true' }
      },
      templateIdentifier: 'manjutesttemplate'
    })
  })
})

describe('<TemplateDetails /> tests', () => {
  beforeAll(() => {
    templateFactory.registerTemplate(new StepTemplate())
  })
  beforeEach(() => {
    jest
      .spyOn(commonHooks, 'useMutateAsGet')
      .mockImplementation(jest.fn().mockReturnValue(mockTemplatesSuccessResponse))
  })

  const baseProps = {
    template: defaultTo(mockTemplates?.data?.content?.[0], {})
  }
  test('should render component properly', async () => {
    const { getByText } = render(
      <TestWrapper>
        <ComponentWrapper {...baseProps} />
      </TestWrapper>
    )

    expect(getByText('manju-test-template-qq-12344')).toBeDefined()
    expect(getByText('templatesLibrary.openInTemplateStudio')).toBeDefined()
    expect(getByText('details')).toBeDefined()
    expect(getByText('activityLog')).toBeDefined()
    expect(getByText('Manual Approval')).toBeDefined()
    expect(getByText('description')).toBeDefined()
    expect(getByText('tagsLabel')).toBeDefined()
    expect(getByText('Internal 1')).toBeDefined()
    expect(getByText('BLUE')).toBeDefined()
    expect(getByText('Tag A')).toBeDefined()
    expect(getByText('pipeline.templateInputs')).toBeDefined()
    expect(getByText('common.yaml')).toBeDefined()
    expect(getByText('templatesLibrary.referencedBy')).toBeDefined()
  })

  test('should render no references when reference by tab is selected', async () => {
    const { getByText, queryByText } = render(
      <TestWrapper>
        <ComponentWrapper {...baseProps} />
      </TestWrapper>
    )

    await act(async () => {
      fireEvent.click(queryByText('templatesLibrary.referencedBy')!)
    })

    expect(getByText('common.noRefData')).toBeDefined()
  })

  test('should show selected version label', async () => {
    const { getByTestId } = render(
      <TestWrapper defaultAppStoreValues={{ isGitSyncEnabled: true }}>
        <ComponentWrapper {...baseProps} isStandAlone />
      </TestWrapper>
    )
    const dropValue = getByTestId('dropdown-value')
    expect(dropValue).toHaveTextContent('v4COMMON.STABLE')
  })

  test('should show always use stable version of the template ', async () => {
    const newBaseProps = produce(baseProps, draft => {
      unset(draft, 'template.versionLabel')
    })
    const { getByTestId } = render(
      <TestWrapper>
        <ComponentWrapper {...newBaseProps} isStandAlone />
      </TestWrapper>
    )
    const dropValue = getByTestId('dropdown-value')
    expect(dropValue).toHaveTextContent('templatesLibrary.alwaysUseStableVersion')
  })

  test('should open template studio on clicking open in template studio', async () => {
    const { getByRole, getByTestId } = render(
      <TestWrapper>
        <ComponentWrapper {...baseProps} />
      </TestWrapper>
    )

    await act(async () => {
      fireEvent.click(getByRole('button', { name: 'templatesLibrary.openInTemplateStudio' }))
    })
    expect(getByTestId('location')).toMatchInlineSnapshot(`
      <div
        data-testid="location"
      >
        /account/kmpySmUISimoRrJL6NL73w/home/orgs/default/projects/Templateproject/setup/resources/template-studio/Step/template/manjutesttemplate/?versionLabel=v4
      </div>
    `)
  })

  test('should match snapshot when error occurs in useMutateAsGet', async () => {
    jest.spyOn(commonHooks, 'useMutateAsGet').mockImplementation(() => {
      return { loading: false, error: 'Some error occurred', data: undefined, refetch: jest.fn() } as any
    })
    const { getByText } = render(
      <TestWrapper>
        <ComponentWrapper {...baseProps} />
      </TestWrapper>
    )
    expect(getByText('We cannot perform your request at the moment. Please try again.')).toBeDefined()
  })
})
